+++ actionpack/lib/action_dispatch/middleware/params_parser.rb
@@ -10,24 +10,8 @@ module ActionDispatch
     # Raised when raw data from the request cannot be parsed by the parser
     # defined for request's content mime type.
     class ParseError < StandardError
-      def initialize(message = nil, original_exception = nil)
-        if message
-          ActiveSupport::Deprecation.warn("Passing #message is deprecated and has no effect. " \
-                                          "#{self.class} will automatically capture the message " \
-                                          "of the original exception.", caller)
-        end
-
-        if original_exception
-          ActiveSupport::Deprecation.warn("Passing #original_exception is deprecated and has no effect. " \
-                                          "Exceptions will automatically capture the original exception.", caller)
-        end
-
+      def initialize
         super($!.message)
-      end
-
-      def original_exception
-        ActiveSupport::Deprecation.warn("#original_exception is deprecated. Use #cause instead.", caller)
-        cause
       end
     end